requirements.txt
streamlit==0.88.0
tensorflow==2.5.1
opencv-python==4.5.2.54
numpy==1.19.5
Pillow==8.3.0



